Create AboutPage API with Journal Structure Field
- Enabled the wagtail 'pages' api which lets you query journal pages
- Added a custom field to AboutPage to get the hierarchical structure of
  the journal

Usage (on local devstack):
- To get a json blob of all pages in the Journals service go to:
  http://localhost:18606/api/v1/content/pages/
  - There are a variety of filtering and ordering options. For more
    information see the wagtail documentation:
    http://docs.wagtail.io/en/v1.7/advanced_topics/api/v2/usage.html
- To get the hierarchical structure of a journal go to:
  http://localhost:18606/api/v1/content/pages/?slug=JOURNAL-ABOUT-PAGE-SLUG&type=journals.JournalAboutPage&fields=structure

Remaining Work:
- Testing for this API endpoint
